@@ -12,7 +12,7 @@
     "lint:fix": "next lint . --fix && prettier . --write",
     "lint:next": "next lint",
     "lint:prettier": "prettier . --check",
-    "lint:stylelint": "stylelint \"**/*.{css,scss}\"",
+    "lint:stylelint": "stylelint \"./src/**/*.{css,scss}\"",
     "lint:tsc": "tsc -p tsconfig.json --noEmit --incremental false",
     "start": "next start"
   },
@@ -21,17 +21,19 @@
     "@blockchain-lab-um/masca-types": "*",
     "@blockchain-lab-um/oidc-types": "*",
     "@blockchain-lab-um/utils": "*",
-    "@headlessui/react": "^1.7.14",
+    "@headlessui/react": "^1.7.15",
     "@heroicons/react": "^2.0.18",
     "@metamask/detect-provider": "^2.0.0",
     "@metamask/providers": "^10.2.0",
     "@radix-ui/react-toast": "^1.1.4",
-    "@tanstack/react-table": "^8.9.1",
+    "@tanstack/react-table": "^8.9.2",
     "@veramo/core": "5.2.0",
+    "@vercel/analytics": "^1.0.1",
     "clsx": "^1.2.1",
-    "ethers": "^6.4.0",
-    "next": "13.4.4",
-    "next-intl": "^2.14.6",
+    "ethers": "^6.5.1",
+    "lint-staged": "^13.2.2",
+    "next": "13.4.5",
+    "next-intl": "2.15.0-beta.5",
     "next-themes": "^0.2.1",
     "prettier-plugin-tailwindcss": "^0.3.0",
     "prop-types": "^15.8.1",
@@ -42,17 +44,17 @@
     "zustand": "^4.3.8"
   },
   "devDependencies": {
-    "@ianvs/prettier-plugin-sort-imports": "^4.0.1",
+    "@ianvs/prettier-plugin-sort-imports": "^4.0.2",
     "@svgr/webpack": "^8.0.1",
-    "@types/eslint": "^8.40.0",
+    "@types/eslint": "^8.40.1",
     "@types/node": "^18.16.15",
     "@types/qs": "^6.9.7",
-    "@types/react": "18.2.7",
+    "@types/react": "18.2.11",
     "@types/react-dom": "18.2.4",
     "autoprefixer": "^10.4.14",
     "cross-env": "^7.0.3",
-    "eslint": "^8.41.0",
-    "eslint-config-next": "^13.4.4",
+    "eslint": "^8.42.0",
+    "eslint-config-next": "^13.4.5",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-unused-imports": "^2.0.0",
@@ -60,8 +62,8 @@
     "prettier": "^2.8.4",
     "prettier-plugin-packagejson": "^2.4.3",
     "rimraf": "^5.0.1",
-    "sass": "^1.62.1",
-    "stylelint": "^15.6.2",
+    "sass": "^1.63.3",
+    "stylelint": "^15.7.0",
     "stylelint-config-css-modules": "^4.2.0",
     "stylelint-config-prettier-scss": "^1.0.0",
     "stylelint-config-standard-scss": "^9.0.0",
