@@ -1,13 +1,14 @@
-import React, { useEffect } from 'react';
-import { useRouter } from 'next/router';
+'use client';
+
+import { useEffect } from 'react';
 import { enableMasca } from '@blockchain-lab-um/masca-connector';
 import { isError } from '@blockchain-lab-um/utils';
 import detectEthereumProvider from '@metamask/detect-provider';
 import { useTranslations } from 'next-intl';
 import { shallow } from 'zustand/shallow';
 
+import Button from '@/components/Button';
 import { useGeneralStore, useMascaStore } from '@/stores';
-import Button from '../Button';
 
 const snapId =
   process.env.USE_LOCAL === 'true'
@@ -62,8 +63,6 @@ const MetaMaskProvider = ({ children }: MetaMaskProviderProps) => {
     }),
     shallow
   );
-
-  const router = useRouter();
 
   const checkMetaMaskCompatibility = async () => {
     try {
@@ -179,39 +178,43 @@ const MetaMaskProvider = ({ children }: MetaMaskProviderProps) => {
     });
   }, [hasMM, hasFlask, address]);
 
-  if (router.pathname === '/' || (hasMM && hasFlask)) {
+  if (hasMM && hasFlask) {
     return <>{children}</>;
   }
 
   return (
-    <div className="flex min-h-full w-full items-center justify-center">
-      <h3 className="text-h3 dark:text-navy-blue-50 text-gray-800">
-        {t('flask')}
-        <div className="mt-16 flex items-center justify-center">
-          <Button
-            variant="gray"
-            onClick={() => {
-              window.open('https://metamask.io/flask/');
-            }}
-          >
-            MetaMask Flask
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              fill="none"
-              viewBox="0 0 24 24"
-              strokeWidth={1.5}
-              stroke="currentColor"
-              className="h-6 w-6"
+    <div className="dark:bg-navy-blue-800 dark:text-navy-blue-400 flex-1 rounded-3xl bg-white shadow-lg">
+      <div className="flex h-full items-center justify-center">
+        <div>
+          <h3 className="text-h4 md:text-h3 dark:text-navy-blue-50 text-gray-800">
+            {t('flask')}
+          </h3>
+          <div className="mt-16 flex items-center justify-center">
+            <Button
+              variant="gray"
+              onClick={() => {
+                window.open('https://metamask.io/flask/');
+              }}
             >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                d="M13.19 8.688a4.5 4.5 0 011.242 7.244l-4.5 4.5a4.5 4.5 0 01-6.364-6.364l1.757-1.757m13.35-.622l1.757-1.757a4.5 4.5 0 00-6.364-6.364l-4.5 4.5a4.5 4.5 0 001.242 7.244"
-              />
-            </svg>
-          </Button>
+              MetaMask Flask
+              <svg
+                xmlns="http://www.w3.org/2000/svg"
+                fill="none"
+                viewBox="0 0 24 24"
+                strokeWidth={1.5}
+                stroke="currentColor"
+                className="h-6 w-6"
+              >
+                <path
+                  strokeLinecap="round"
+                  strokeLinejoin="round"
+                  d="M13.19 8.688a4.5 4.5 0 011.242 7.244l-4.5 4.5a4.5 4.5 0 01-6.364-6.364l1.757-1.757m13.35-.622l1.757-1.757a4.5 4.5 0 00-6.364-6.364l-4.5 4.5a4.5 4.5 0 001.242 7.244"
+                />
+              </svg>
+            </Button>
+          </div>
         </div>
-      </h3>
+      </div>
     </div>
   );
 };
