@@ -1,6 +1,8 @@
-import React, { useEffect, useState } from 'react';
+'use client';
+
+import { useEffect, useState } from 'react';
 import Link from 'next/link';
-import { useRouter } from 'next/router';
+import { useRouter } from 'next/navigation';
 import { QueryVCsRequestResult } from '@blockchain-lab-um/masca-types';
 import { isError } from '@blockchain-lab-um/utils';
 import {
@@ -28,6 +30,7 @@ import {
   SortingState,
   useReactTable,
 } from '@tanstack/react-table';
+import clsx from 'clsx';
 import { useTranslations } from 'next-intl';
 import { shallow } from 'zustand/shallow';
 
@@ -428,34 +431,27 @@ const Table = () => {
                           cell.column.id !== 'issuer' &&
                           cell.column.id !== 'actions'
                         ) {
-                          router
-                            .push(
-                              {
-                                pathname: '/verifiable-credential',
-                                query: { id: row.original.metadata.id },
-                              },
-                              undefined,
-                              { shallow: true }
-                            )
-                            .then(() => {})
-                            .catch(() => {});
+                          router.push(
+                            `/app/verifiable-credential/${row.original.metadata.id}`
+                          );
                         }
                       }}
-                      className={`max-h-16 py-5  ${
+                      className={clsx(
+                        'max-h-16 py-5',
                         cell.column.id === 'exp_date' ||
-                        cell.column.id === 'date'
+                          cell.column.id === 'date'
                           ? 'hidden lg:table-cell'
+                          : '',
+                        cell.column.id === 'type'
+                          ? 'w-[20%] max-w-[20%] px-2'
+                          : '',
+                        cell.column.id === 'actions'
+                          ? 'hidden sm:table-cell'
+                          : '',
+                        cell.column.id === 'subject'
+                          ? 'hidden xl:table-cell'
                           : ''
-                      }
-                        ${
-                          cell.column.id === 'type'
-                            ? ' w-[20%] max-w-[20%] px-2'
-                            : ''
-                        }
-                  ${cell.column.id === 'actions' ? 'hidden sm:table-cell' : ''}
-                  ${
-                    cell.column.id === 'subject' ? 'hidden xl:table-cell' : ''
-                  }`}
+                      )}
                       key={cell.id}
                     >
                       {flexRender(
@@ -468,12 +464,12 @@ const Table = () => {
               ))}
             </tbody>
           </table>
-          <div className=" mt-auto flex justify-center rounded-b-3xl pb-3 pt-3">
+          <div className="mt-auto flex justify-center rounded-b-3xl pb-3 pt-3">
             <TablePagination table={table} />
           </div>
           {table.getSelectedRowModel().rows.length > 0 && (
-            <div className="mb-2 max-lg:flex max-lg:justify-center lg:absolute lg:-bottom-5 lg:right-10">
-              <Link href="create-verifiable-presentation">
+            <div className="absolute -bottom-3 right-10 md:-bottom-4 lg:-bottom-5">
+              <Link href="/app/create-verifiable-presentation">
                 <Button
                   variant="primary"
                   size="wd"
@@ -500,50 +496,48 @@ const Table = () => {
     );
   }
   return (
-    <>
-      <div className="relative flex h-full min-h-[50vh] w-full flex-col">
-        <div className="dark:border-navy-blue-600 flex items-center justify-between border-b border-gray-400 p-5">
-          <div className="text-h2 font-ubuntu dark:text-navy-blue-50 pl-4 font-medium text-gray-900">
-            {t('table-header.credentials')}
+    <div className="relative flex h-full min-h-[50vh] w-full flex-col">
+      <div className="dark:border-navy-blue-600 flex items-center justify-between border-b border-gray-400 p-5">
+        <div className="text-h2 font-ubuntu dark:text-navy-blue-50 pl-4 font-medium text-gray-900">
+          {t('table-header.credentials')}
+        </div>
+        <div className="text-right">
+          <div className="text-h4 dark:text-navy-blue-50 text-gray-900">
+            {vcs.length} {t('table-header.found')}
           </div>
-          <div className="text-right">
-            <div className="text-h4 dark:text-navy-blue-50 text-gray-900">
-              {vcs.length} {t('table-header.found')}
-            </div>
-            <div className="text-h5 dark:text-navy-blue-400 text-gray-600">
-              {t('table-header.fetched')}: today
-            </div>
+          <div className="text-h5 dark:text-navy-blue-400 text-gray-600">
+            {t('table-header.fetched')}: today
           </div>
         </div>
-        <div className="flex flex-wrap justify-center">
-          {table.getRowModel().rows.map((row, key) => (
-            <VCCard key={key} row={row} />
-          ))}
-        </div>
-        <div className="mt-auto flex justify-center rounded-b-3xl pb-3 pt-3">
-          <TablePagination table={table} />
-        </div>
-        {table.getSelectedRowModel().rows.length > 0 && (
-          <div className="absolute -bottom-5 right-10">
-            <Link href="create-verifiable-presentation">
-              <Button
-                variant="primary"
-                size="wd"
-                onClick={() => {
-                  setSelectedVCs(
-                    table.getSelectedRowModel().rows.map((r) => r.original)
-                  );
-                }}
-              >
-                {t('create-verifiable-presentation')}{' '}
-                {table.getSelectedRowModel().rows.length > 0 &&
-                  `(${table.getSelectedRowModel().rows.length})`}
-              </Button>
-            </Link>
-          </div>
-        )}
       </div>
-    </>
+      <div className="flex flex-wrap justify-center">
+        {table.getRowModel().rows.map((row, key) => (
+          <VCCard key={key} row={row} />
+        ))}
+      </div>
+      <div className="mt-auto flex justify-center rounded-b-3xl pb-3 pt-3">
+        <TablePagination table={table} />
+      </div>
+      {table.getSelectedRowModel().rows.length > 0 && (
+        <div className="absolute -bottom-3 right-10 md:-bottom-4 lg:-bottom-5">
+          <Link href="/app/create-verifiable-presentation">
+            <Button
+              variant="primary"
+              size="wd"
+              onClick={() => {
+                setSelectedVCs(
+                  table.getSelectedRowModel().rows.map((r) => r.original)
+                );
+              }}
+            >
+              {t('create-verifiable-presentation')}{' '}
+              {table.getSelectedRowModel().rows.length > 0 &&
+                `(${table.getSelectedRowModel().rows.length})`}
+            </Button>
+          </Link>
+        </div>
+      )}
+    </div>
   );
 };
 
