@@ -7,6 +7,7 @@ import {
   type TokenRequest,
   type TokenResponse,
 } from '@blockchain-lab-um/oidc-types';
+import { qsCustomDecoder } from '@blockchain-lab-um/utils';
 import { HttpServer } from '@nestjs/common';
 import {
   FastifyAdapter,
@@ -43,8 +44,13 @@ const credOfferAndTokenRequest = async (server: HttpServer<any, any>) => {
     .send();
 
   const query = qs.parse(
-    response.text.replace('openid-credential-offer://?', '')
-  ).credential_offer as unknown as CredentialOffer;
+    response.text.replace('openid-credential-offer://?', ''),
+    {
+      depth: 50,
+      parameterLimit: 1000,
+      decoder: qsCustomDecoder,
+    }
+  ).credential_offer as CredentialOffer;
 
   expect(response.status).toBe(200);
   const tokenRequestData: TokenRequest = {
@@ -201,14 +207,20 @@ describe('Issuer controller', () => {
         expect(response.status).toBe(200);
 
         const query = qs.parse(
-          response.text.replace('openid-credential-offer://?', '')
-        ).credential_offer as unknown as CredentialOffer;
+          response.text.replace('openid-credential-offer://?', ''),
+          {
+            depth: 50,
+            parameterLimit: 1000,
+            decoder: qsCustomDecoder,
+          }
+        ).credential_offer as CredentialOffer;
 
         expect(query.credentials).toStrictEqual(['GmCredential']);
         expect(query.credential_issuer).toBeDefined();
         expect(query.grants).toStrictEqual({
           'urn:ietf:params:oauth:grant-type:pre-authorized_code': {
             'pre-authorized_code': expect.any(String),
+            user_pin_required: false,
           },
         });
 
@@ -237,8 +249,13 @@ describe('Issuer controller', () => {
         expect(response.status).toBe(200);
 
         const query = qs.parse(
-          response.text.replace('openid-credential-offer://?', '')
-        ).credential_offer as unknown as CredentialOffer;
+          response.text.replace('openid-credential-offer://?', ''),
+          {
+            depth: 50,
+            parameterLimit: 1000,
+            decoder: qsCustomDecoder,
+          }
+        ).credential_offer as CredentialOffer;
 
         expect(query.credentials).toStrictEqual([
           {
@@ -250,6 +267,7 @@ describe('Issuer controller', () => {
         expect(query.grants).toStrictEqual({
           'urn:ietf:params:oauth:grant-type:pre-authorized_code': {
             'pre-authorized_code': expect.any(String),
+            user_pin_required: false,
           },
         });
 
@@ -278,8 +296,13 @@ describe('Issuer controller', () => {
         expect(response.status).toBe(200);
 
         const query = qs.parse(
-          response.text.replace('openid-credential-offer://?', '')
-        ).credential_offer as unknown as CredentialOffer;
+          response.text.replace('openid-credential-offer://?', ''),
+          {
+            depth: 50,
+            parameterLimit: 1000,
+            decoder: qsCustomDecoder,
+          }
+        ).credential_offer as CredentialOffer;
 
         expect(query.credentials).toStrictEqual([
           {
@@ -291,6 +314,7 @@ describe('Issuer controller', () => {
         expect(query.grants).toStrictEqual({
           'urn:ietf:params:oauth:grant-type:pre-authorized_code': {
             'pre-authorized_code': expect.any(String),
+            user_pin_required: false,
           },
         });
 
@@ -319,8 +343,13 @@ describe('Issuer controller', () => {
         expect(response.status).toBe(200);
 
         const query = qs.parse(
-          response.text.replace('openid-credential-offer://?', '')
-        ).credential_offer as unknown as CredentialOffer;
+          response.text.replace('openid-credential-offer://?', ''),
+          {
+            depth: 50,
+            parameterLimit: 1000,
+            decoder: qsCustomDecoder,
+          }
+        ).credential_offer as CredentialOffer;
 
         expect(query.credentials).toStrictEqual([
           {
@@ -332,6 +361,7 @@ describe('Issuer controller', () => {
         expect(query.grants).toStrictEqual({
           'urn:ietf:params:oauth:grant-type:pre-authorized_code': {
             'pre-authorized_code': expect.any(String),
+            user_pin_required: false,
           },
         });
 
@@ -372,8 +402,13 @@ describe('Issuer controller', () => {
         expect(response.status).toBe(200);
 
         const query = qs.parse(
-          response.text.replace('openid-credential-offer://?', '')
-        ).credential_offer as unknown as CredentialOffer;
+          response.text.replace('openid-credential-offer://?', ''),
+          {
+            depth: 50,
+            parameterLimit: 1000,
+            decoder: qsCustomDecoder,
+          }
+        ).credential_offer as CredentialOffer;
 
         expect(query.credentials).toStrictEqual([
           {
@@ -397,6 +432,7 @@ describe('Issuer controller', () => {
         expect(query.grants).toStrictEqual({
           'urn:ietf:params:oauth:grant-type:pre-authorized_code': {
             'pre-authorized_code': expect.any(String),
+            user_pin_required: false,
           },
         });
 
@@ -425,8 +461,13 @@ describe('Issuer controller', () => {
         expect(response.status).toBe(200);
 
         const query = qs.parse(
-          response.text.replace('openid-credential-offer://?', '')
-        ).credential_offer as unknown as CredentialOffer;
+          response.text.replace('openid-credential-offer://?', ''),
+          {
+            depth: 50,
+            parameterLimit: 1000,
+            decoder: qsCustomDecoder,
+          }
+        ).credential_offer as CredentialOffer;
 
         expect(query.credentials).toStrictEqual([
           {
@@ -438,6 +479,7 @@ describe('Issuer controller', () => {
         expect(query.grants).toStrictEqual({
           'urn:ietf:params:oauth:grant-type:pre-authorized_code': {
             'pre-authorized_code': expect.any(String),
+            user_pin_required: false,
           },
         });
 
@@ -467,8 +509,13 @@ describe('Issuer controller', () => {
         expect(response.status).toBe(200);
 
         const query = qs.parse(
-          response.text.replace('openid-credential-offer://?', '')
-        ).credential_offer as unknown as CredentialOffer;
+          response.text.replace('openid-credential-offer://?', ''),
+          {
+            depth: 50,
+            parameterLimit: 1000,
+            decoder: qsCustomDecoder,
+          }
+        ).credential_offer as CredentialOffer;
 
         expect(query.credentials).toStrictEqual([
           {
@@ -480,7 +527,7 @@ describe('Issuer controller', () => {
         expect(query.grants).toStrictEqual({
           'urn:ietf:params:oauth:grant-type:pre-authorized_code': {
             'pre-authorized_code': expect.any(String),
-            user_pin_required: 'false',
+            user_pin_required: false,
           },
         });
 
@@ -510,8 +557,13 @@ describe('Issuer controller', () => {
         expect(response.status).toBe(200);
 
         const query = qs.parse(
-          response.text.replace('openid-credential-offer://?', '')
-        ).credential_offer as unknown as CredentialOffer;
+          response.text.replace('openid-credential-offer://?', ''),
+          {
+            depth: 50,
+            parameterLimit: 1000,
+            decoder: qsCustomDecoder,
+          }
+        ).credential_offer as CredentialOffer;
 
         expect(query.credentials).toStrictEqual([
           {
@@ -523,7 +575,7 @@ describe('Issuer controller', () => {
         expect(query.grants).toStrictEqual({
           'urn:ietf:params:oauth:grant-type:pre-authorized_code': {
             'pre-authorized_code': expect.any(String),
-            user_pin_required: 'true',
+            user_pin_required: true,
           },
         });
 
@@ -552,8 +604,13 @@ describe('Issuer controller', () => {
         expect(response.status).toBe(200);
 
         const query = qs.parse(
-          response.text.replace('openid-credential-offer://?', '')
-        ).credential_offer as unknown as CredentialOffer;
+          response.text.replace('openid-credential-offer://?', ''),
+          {
+            depth: 50,
+            parameterLimit: 1000,
+            decoder: qsCustomDecoder,
+          }
+        ).credential_offer as CredentialOffer;
 
         expect(query.credentials).toStrictEqual([
           {
@@ -597,8 +654,13 @@ describe('Issuer controller', () => {
         expect(response.status).toBe(200);
 
         const query = qs.parse(
-          response.text.replace('openid-credential-offer://?', '')
-        ).credential_offer as unknown as CredentialOffer;
+          response.text.replace('openid-credential-offer://?', ''),
+          {
+            depth: 50,
+            parameterLimit: 1000,
+            decoder: qsCustomDecoder,
+          }
+        ).credential_offer as CredentialOffer;
 
         expect(query.credentials).toStrictEqual([
           {
@@ -613,7 +675,7 @@ describe('Issuer controller', () => {
           },
           'urn:ietf:params:oauth:grant-type:pre-authorized_code': {
             'pre-authorized_code': expect.any(String),
-            user_pin_required: 'true',
+            user_pin_required: true,
           },
         });
 
@@ -683,8 +745,13 @@ describe('Issuer controller', () => {
           .send();
 
         const query = qs.parse(
-          response.text.replace('openid-credential-offer://?', '')
-        ).credential_offer as unknown as CredentialOffer;
+          response.text.replace('openid-credential-offer://?', ''),
+          {
+            depth: 50,
+            parameterLimit: 1000,
+            decoder: qsCustomDecoder,
+          }
+        ).credential_offer as CredentialOffer;
 
         expect(response.status).toBe(200);
 
@@ -735,8 +802,13 @@ describe('Issuer controller', () => {
           .send();
 
         const query = qs.parse(
-          response.text.replace('openid-credential-offer://?', '')
-        ).credential_offer as unknown as CredentialOffer;
+          response.text.replace('openid-credential-offer://?', ''),
+          {
+            depth: 50,
+            parameterLimit: 1000,
+            decoder: qsCustomDecoder,
+          }
+        ).credential_offer as CredentialOffer;
 
         expect(response.status).toBe(200);
 
@@ -841,8 +913,13 @@ describe('Issuer controller', () => {
         expect(response.status).toBe(200);
 
         const query = qs.parse(
-          response.text.replace('openid-credential-offer://?', '')
-        ).credential_offer as unknown as CredentialOffer;
+          response.text.replace('openid-credential-offer://?', ''),
+          {
+            depth: 50,
+            parameterLimit: 1000,
+            decoder: qsCustomDecoder,
+          }
+        ).credential_offer as CredentialOffer;
 
         const tokenRequestData: TokenRequest = {
           grant_type: 'urn:ietf:params:oauth:grant-type:pre-authorized_code',
@@ -935,8 +1012,13 @@ describe('Issuer controller', () => {
         expect(response.status).toBe(200);
 
         const query = qs.parse(
-          response.text.replace('openid-credential-offer://?', '')
-        ).credential_offer as unknown as CredentialOffer;
+          response.text.replace('openid-credential-offer://?', ''),
+          {
+            depth: 50,
+            parameterLimit: 1000,
+            decoder: qsCustomDecoder,
+          }
+        ).credential_offer as CredentialOffer;
 
         const tokenRequestData: TokenRequest = {
           grant_type: 'urn:ietf:params:oauth:grant-type:pre-authorized_code',
