@@ -1,8 +1,11 @@
+'use client';
+
 import { Fragment } from 'react';
 import type { AvailableMethods } from '@blockchain-lab-um/masca-types';
 import { isError } from '@blockchain-lab-um/utils';
 import { Menu, Transition } from '@headlessui/react';
 import { ChevronDownIcon } from '@heroicons/react/20/solid';
+import clsx from 'clsx';
 import { shallow } from 'zustand/shallow';
 
 import { useMascaStore, useToastStore } from '@/stores';
@@ -53,22 +56,24 @@ export default function MethodDropdownMenu() {
   };
 
   return (
-    <Menu as="div" className="relative z-20 mx-2">
+    <Menu as="div" className="relative z-20 mx-2 hidden md:block">
       {({ open }) => (
         <Fragment>
           <div>
             <Menu.Button
-              className={`dark:text-navy-blue-400 text-h5 font-ubuntu animated-transition inline-flex w-full justify-center rounded-3xl px-4 py-2 font-thin text-gray-600 focus:outline-none ${
+              className={clsx(
+                'dark:text-navy-blue-400 text-h5 font-ubuntu animated-transition inline-flex w-full justify-center rounded-3xl px-4 py-2 font-thin text-gray-600 outline-none focus:outline-none focus-visible:outline-none',
                 open
                   ? 'dark:bg-navy-blue-800 bg-orange-100/50'
                   : 'dark:hover:bg-navy-blue-800 hover:bg-orange-100/50'
-              }`}
+              )}
             >
               {currMethod}
               <ChevronDownIcon
-                className={`dark:text-navy-blue-400 animated-transition -mr-1 ml-2 h-5 w-5 text-gray-600 max-md:rotate-180 ${
+                className={clsx(
+                  'dark:text-navy-blue-400 animated-transition -mr-1 ml-2 h-5 w-5 text-gray-600',
                   open ? 'rotate-180' : ''
-                }`}
+                )}
               />
             </Menu.Button>
           </div>
@@ -76,13 +81,13 @@ export default function MethodDropdownMenu() {
           <Transition
             show={open}
             enter="transition ease-out duration-100"
-            enterFrom="transform opacity-0 scale-95"
-            enterTo="transform opacity-100 scale-100"
+            enterFrom="opacity-0 scale-95"
+            enterTo="opacity-100 scale-100"
             leave="transition ease-in duration-75"
-            leaveFrom="transform opacity-100 scale-100"
-            leaveTo="transform opacity-0 scale-95"
+            leaveFrom="opacity-100 scale-100"
+            leaveTo="opacity-0 scale-95"
           >
-            <Menu.Items className="dark:bg-navy-blue-500 absolute right-0 mt-1 w-48 rounded-3xl bg-white shadow-lg max-md:-top-12 max-md:-translate-y-full max-md:transform">
+            <Menu.Items className="dark:bg-navy-blue-500 absolute right-0 mt-1 w-48 rounded-3xl bg-white shadow-lg">
               <div className="p-1 text-center ">
                 {methods.map((method, id) => (
                   <DropdownButton
