@@ -84,7 +84,6 @@ function ModifyDSModal({ open, setOpen, vc }: ModifyDSModalProps) {
     setType('normal');
     setTitle('Saving Credential');
     setToastOpen(true);
-    setOpen(false);
 
     const res = await api.saveVC(vc.data, { store });
 
@@ -131,7 +130,7 @@ function ModifyDSModal({ open, setOpen, vc }: ModifyDSModalProps) {
           leaveFrom="opacity-100"
           leaveTo="opacity-0"
         >
-          <div className="fixed inset-0 bg-black bg-opacity-25 dark:bg-opacity-60" />
+          <div className="fixed inset-0 bg-black/25 dark:bg-black/60" />
         </Transition.Child>
 
         <div className="fixed inset-0 overflow-y-auto">
@@ -145,7 +144,7 @@ function ModifyDSModal({ open, setOpen, vc }: ModifyDSModalProps) {
               leaveFrom="opacity-100 scale-100"
               leaveTo="opacity-0 scale-95"
             >
-              <Dialog.Panel className="dark:bg-navy-blue-500 w-full max-w-md transform overflow-hidden rounded-2xl bg-white p-6 text-left align-middle shadow-xl transition-all">
+              <Dialog.Panel className="dark:bg-navy-blue-500 w-full max-w-md overflow-hidden rounded-2xl bg-white p-6 text-left align-middle shadow-xl transition-all">
                 <Dialog.Title
                   as="h3"
                   className="font-ubuntu dark:text-navy-blue-50 text-xl font-medium leading-6 text-gray-900"
