@@ -1,4 +1,3 @@
-import React from 'react';
 import Image from 'next/image';
 
 type StoreIconProps = {
@@ -6,7 +5,7 @@ type StoreIconProps = {
 };
 
 const logo: Record<string, string> = {
-  snap: 'ssi_icon_b.png',
+  snap: 'masca_black.png',
   ceramic: 'ceramic_logo.png',
 };
 
