@@ -1,5 +1,8 @@
+'use client';
+
 import { useEffect, useState } from 'react';
 import { MoonIcon, SunIcon } from '@heroicons/react/24/outline';
+import clsx from 'clsx';
 import { useTheme } from 'next-themes';
 
 const ToggleTheme = () => {
@@ -18,9 +21,11 @@ const ToggleTheme = () => {
 
   return (
     <button
-      className={`animated-transition dark:hover:bg-navy-blue-800 dark:text-navy-blue-400 m-2 flex h-[32px] w-[32px] items-center justify-center rounded-full p-1 hover:bg-orange-100 lg:h-[38px] lg:w-[38px] ${
+      className={clsx(
+        'animated-transition dark:hover:bg-navy-blue-800 dark:text-navy-blue-400 m-2 flex h-[32px] w-[32px] items-center justify-center rounded-full p-1 hover:bg-orange-100 lg:h-[38px] lg:w-[38px]',
+        'outline-none focus-visible:outline-none',
         resolvedTheme === 'dark' ? 'text-white/50' : 'text-black/50'
-      }`}
+      )}
       type="button"
       onClick={toggleTheme}
     >
