@@ -1,4 +1,4 @@
-import React, { Fragment, useEffect, useState } from 'react';
+import { Fragment, useEffect, useState } from 'react';
 import { QueryVCsRequestResult } from '@blockchain-lab-um/masca-types';
 import { Combobox, Transition } from '@headlessui/react';
 import { CheckIcon, ChevronDownIcon } from '@heroicons/react/20/solid';
@@ -36,6 +36,16 @@ const DataStoreCombobox = ({ vcs, isConnected }: DataStoreComboboxProps) => {
   }) as string[];
 
   useEffect(() => {
+    // If there is only ceramic, select it (snap is selected by default)
+    if (
+      dataStores.length === 1 &&
+      dataStores.includes('ceramic') &&
+      !selectedItems.includes('ceramic')
+    ) {
+      setColumnFilters([{ id: 'data_store', value: ['ceramic'] }]);
+      selectedItems.push('ceramic');
+    }
+
     // inner join selectedItems and dataStores
     const filteredDataStores = dataStores.filter((ds) =>
       selectedItems.includes(ds)
@@ -63,9 +73,9 @@ const DataStoreCombobox = ({ vcs, isConnected }: DataStoreComboboxProps) => {
       >
         {({ open }) => (
           <div className="relative">
-            <div className="w-34 relative cursor-default overflow-hidden rounded-full shadow-md sm:text-sm">
+            <div className="md:w-34 relative h-[37px] w-24 cursor-default overflow-hidden rounded-full shadow-md sm:text-sm md:h-[43px]">
               <Combobox.Input
-                className={`text-md dark:bg-navy-blue-700 dark:text-navy-blue-50 py-3 pl-5 text-gray-700 focus:outline-none ${
+                className={`md:text-md dark:bg-navy-blue-700 dark:text-navy-blue-50 py-2.5 pl-5 text-sm text-gray-700 focus:outline-none md:py-3 ${
                   !isConnected || vcs.length === 0
                     ? 'bg-gray-50 text-gray-300'
                     : ' '
@@ -81,7 +91,7 @@ const DataStoreCombobox = ({ vcs, isConnected }: DataStoreComboboxProps) => {
               <Combobox.Button className="absolute inset-y-0 right-0 flex items-center pr-2">
                 <>
                   <ChevronDownIcon
-                    className={`animated-transition dark:text-navy-blue-50 h-5 w-5 text-gray-700 ${
+                    className={`animated-transition dark:text-navy-blue-50 h-3 w-3 text-gray-700 md:h-4 md:w-4 lg:h-5 lg:w-5 ${
                       !isConnected || vcs.length === 0 ? 'text-gray-300' : ' '
                     } ${open ? 'rotate-180' : ''}`}
                   />
@@ -96,7 +106,7 @@ const DataStoreCombobox = ({ vcs, isConnected }: DataStoreComboboxProps) => {
               leaveTo="opacity-0"
               afterLeave={() => setQuery('')}
             >
-              <Combobox.Options className="dark:bg-navy-blue-600 absolute right-0 z-50 mt-1 w-48 rounded-3xl bg-white p-1 text-center shadow-lg">
+              <Combobox.Options className="dark:bg-navy-blue-600 absolute left-0 z-50 mt-1 w-36 rounded-3xl bg-white p-0.5 text-center shadow-lg md:right-0 md:w-48 md:p-1">
                 {filteredDataStores.map((data_store, id) => (
                   <Combobox.Option key={id} value={data_store}>
                     {({ selected, active }) => (
@@ -108,13 +118,13 @@ const DataStoreCombobox = ({ vcs, isConnected }: DataStoreComboboxProps) => {
                           selected
                             ? 'dark:text-orange-accent-dark dark:bg-navy-blue-600 bg-white text-pink-700'
                             : 'dark:text-navy-blue-100 text-gray-600',
-                          'block rounded-full py-2 text-lg'
+                          'text-md block rounded-full py-2 md:text-lg'
                         )}
                       >
-                        <span className="grid grid-cols-3">
+                        <span className="grid grid-cols-3 items-center">
                           <span>
                             {selected ? (
-                              <CheckIcon className="ml-3 h-5 w-5" />
+                              <CheckIcon className="ml-3 h-4 w-4 md:h-5 md:w-5" />
                             ) : (
                               ''
                             )}
