@@ -1,4 +1,3 @@
-import React from 'react';
 import { QueryVCsRequestResult } from '@blockchain-lab-um/masca-types';
 import { MagnifyingGlassIcon } from '@heroicons/react/20/solid';
 import { useTranslations } from 'next-intl';
@@ -22,7 +21,7 @@ const GlobalFilter = ({ isConnected, vcs }: GlobalFilterProps) => {
   );
   return (
     <div
-      className={`dark:bg-navy-blue-700 dark:text-navy-blue-50 flex flex-1 items-center rounded-full shadow-md sm:flex-none ${
+      className={`dark:bg-navy-blue-700 dark:text-navy-blue-50 flex h-[37px] flex-1 items-center rounded-full shadow-md sm:flex-none md:h-[43px] ${
         !isConnected || vcs.length === 0 ? 'bg-gray-50' : 'bg-white'
       }`}
     >
@@ -37,7 +36,7 @@ const GlobalFilter = ({ isConnected, vcs }: GlobalFilterProps) => {
         onChange={(e) => {
           setGlobalFilter(e.target.value);
         }}
-        className={`font-lg dark:bg-navy-blue-700 dark:text-navy-blue-50 dark:placeholder:text-navy-blue-500 h-full w-full rounded-full pl-3 pr-2 text-gray-700 placeholder:text-gray-400 focus:outline-none sm:px-2 ${
+        className={`text-md dark:bg-navy-blue-700 dark:text-navy-blue-50 dark:placeholder:text-navy-blue-500 h-full w-full rounded-full pl-3 pr-2 text-gray-700 placeholder:text-gray-400 focus:outline-none sm:px-2 md:text-lg ${
           !isConnected || vcs.length === 0
             ? 'bg-gray-50 text-gray-300 placeholder:text-gray-300'
             : 'bg-white'
