@@ -1,7 +1,7 @@
 import type { SignArgs } from '@blockchain-lab-um/oidc-client-plugin';
 import { bytesToBase64url, encodeBase64url } from '@veramo/utils';
 import elliptic from 'elliptic';
-import { sha256 } from 'ethers';
+import { sha256 } from 'ethereum-cryptography/sha256';
 
 const { ec: EC } = elliptic;
 
