@@ -1,3 +1,4 @@
+import { uint8ArrayToHex } from '@blockchain-lab-um/utils';
 import {
   AbstractDataStore,
   type IFilterArgs,
@@ -7,7 +8,7 @@ import { DIDDataStore } from '@glazed/did-datastore';
 import { MetaMaskInpageProvider } from '@metamask/providers';
 import type { SnapsGlobalObject } from '@metamask/snaps-types';
 import type { W3CVerifiableCredential } from '@veramo/core';
-import { sha256 } from 'ethers';
+import { sha256 } from 'ethereum-cryptography/sha256.js';
 import jsonpath from 'jsonpath';
 
 import { aliases, getCeramic } from '../../../utils/ceramicUtils';
@@ -121,7 +122,7 @@ export class CeramicVCStore extends AbstractDataStore {
       'StoredCredentials'
     )) as StoredCredentials;
     if (storedCredentials && storedCredentials.vcs) {
-      const id = sha256(Buffer.from(JSON.stringify(vc))).toString();
+      const id = uint8ArrayToHex(sha256(Buffer.from(JSON.stringify(vc))));
 
       if (storedCredentials.vcs[id]) {
         return id;
@@ -131,7 +132,7 @@ export class CeramicVCStore extends AbstractDataStore {
       await datastore.merge('StoredCredentials', storedCredentials);
       return id;
     }
-    const id = sha256(Buffer.from(JSON.stringify(vc))).toString();
+    const id = uint8ArrayToHex(sha256(Buffer.from(JSON.stringify(vc))));
     const storedCredentialsNew: StoredCredentials = { vcs: {} };
     storedCredentialsNew.vcs[id] = vc;
     await datastore.merge('StoredCredentials', storedCredentialsNew);
